Added look command.
The game displays the description of a thing if the thing is in the same location as the player or in the players inventory. If the player only writes look the game displays the information about the current location. Also added a method to show the locations information without clearing the consol.